refactor: 리다이렉트URL 함수 분리
