+++ src/app/(user)/error.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default function UserError() {
+  return <h1>유저 페이지 서버 에러</h1>;
+}
